+++ src/components/index.ts
@@ -4,5 +4,7 @@ export * from './InfiniteScroll';
 export * from './ItemSelector';
 export * from './Notification';
 export * from './Notifications';
+export * from './PlaceholderContainer';
+export * from './ActionsPanel';
 
 export {Lang, configure} from './utils/configure';